+++ app/src/main/java/com/test/bankingapp/transaction/presentation/screens/AllTransactionsScreen.kt
@@ -69,7 +69,7 @@ fun AllTransactionsScreen(navController: NavController) {
 
     ModalBottomSheetLayout(
         sheetContent = {
-            BottomSheetContent(onDismissRequest = {
+            TransactionBottomSheetContent(onDismissRequest = {
                 coroutineScope.launch {
                     bottomSheetState.hide()
                     isSheetVisible = false
@@ -125,149 +125,6 @@ fun AllTransactionsScreen(navController: NavController) {
             Spacer(modifier = Modifier.height(16.dp))
             RoundedLazyColumn(navController = navController, transactions = transactions)
         }
-    }
-}
-
-
-
-@Composable
-fun BottomSheetContent(
-    onDismissRequest: () -> Unit
-) {
-    var showStartDateCalendarDialog by remember { mutableStateOf(false) }
-    var showEndDateCalendarDialog by remember { mutableStateOf(false) }
-    var startDate by remember { mutableStateOf("") }
-    var endDate by remember { mutableStateOf("") }
-
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .background(colorResource(id = R.color.black))
-            .padding(16.dp)
-    ) {
-        Divider(
-            thickness = 6.dp, color = colorResource(id = R.color.dark_gray),
-            modifier = Modifier
-                .width(45.dp)
-                .clip(RoundedCornerShape(20.dp))
-                .align(Alignment.CenterHorizontally)
-        )
-        Spacer(modifier = Modifier.height(16.dp))
-        Text(
-            text = stringResource(id = R.string.filter_by_date),
-            color = colorResource(id = R.color.white),
-            fontSize = 28.sp,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier.padding(bottom = 16.dp, start = 8.dp)
-        )
-
-        Text(
-            text = stringResource(id = R.string.start_date),
-            color = colorResource(id = R.color.white),
-            fontSize = 18.sp,
-            modifier = Modifier.padding(bottom = 8.dp)
-        )
-        OutlinedTextField(
-            value = startDate,
-            onValueChange = { startDate = it },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(bottom = 8.dp)
-                .height(54.dp)
-                .clickable {
-                    showStartDateCalendarDialog = true
-                },
-            enabled = false,
-            colors = TextFieldDefaults.outlinedTextFieldColors(
-                textColor = Color.White,
-                disabledTextColor = Color.White,
-                disabledBorderColor = Color.White,
-                disabledLabelColor = Color.Gray,
-                cursorColor = colorResource(id = R.color.white)
-            ),
-            keyboardOptions = KeyboardOptions.Default.copy(
-                keyboardType = KeyboardType.Number,
-                imeAction = ImeAction.Next
-            ),
-            shape = RoundedCornerShape(10.dp)
-        )
-
-        if (showStartDateCalendarDialog) {
-            CustomDatePickerDialog(
-                selectedDate = LocalDate.now(),
-                onDateSelected = {
-                    startDate = it.toString()
-                    showStartDateCalendarDialog = false
-                }
-            )
-        }
-
-        Text(
-            text = stringResource(id = R.string.end_date),
-            color = colorResource(id = R.color.white),
-            fontSize = 18.sp,
-            modifier = Modifier.padding(bottom = 8.dp)
-        )
-        OutlinedTextField(
-            value = endDate,
-            onValueChange = { endDate = it },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(bottom = 8.dp)
-                .height(54.dp)
-                .clickable {
-                    showEndDateCalendarDialog = true
-                },
-            enabled = false,
-            colors = TextFieldDefaults.outlinedTextFieldColors(
-                textColor = Color.White,
-                disabledTextColor = Color.White,
-                disabledBorderColor = Color.White,
-                disabledLabelColor = Color.Gray,
-                cursorColor = colorResource(id = R.color.white)
-            ),
-            keyboardOptions = KeyboardOptions.Default.copy(
-                keyboardType = KeyboardType.Number,
-                imeAction = ImeAction.Next
-            ),
-            shape = RoundedCornerShape(10.dp)
-        )
-
-        if (showEndDateCalendarDialog) {
-            CustomDatePickerDialog(
-                selectedDate = LocalDate.now(),
-                onDateSelected = {
-                    endDate = it.toString()
-                    showEndDateCalendarDialog = false
-                }
-            )
-        }
-
-        Spacer(modifier = Modifier.height(16.dp))
-        Divider(
-            thickness = 2.dp, color = colorResource(id = R.color.dark_gray),
-            modifier = Modifier
-                .fillMaxWidth()
-                .clip(RoundedCornerShape(20.dp))
-                .align(Alignment.CenterHorizontally)
-        )
-        Spacer(modifier = Modifier.height(10.dp))
-
-        Button(
-            onClick = { /* TODO */ },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp)
-                .height(50.dp),
-            colors = ButtonDefaults.buttonColors(
-                backgroundColor = colorResource(id = R.color.blue),
-                contentColor = colorResource(id = R.color.white)
-            ),
-            shape = RoundedCornerShape(10.dp)
-        ) {
-            Text(text = stringResource(id = R.string.submit), fontSize = 16.sp)
-        }
-        
     }
 }
